services/chatsentry/requirements.txt
fastapi>=0.135.0
starlette>=0.41.0,<1.0.0
uvicorn>=0.42.0
minio>=7.2.0
psycopg2-binary>=2.9.0
pandas>=3.0.0
python-dotenv>=1.2.0
pyyaml>=6.0.0
ftfy>=6.3.1
emoji>=2.15.0
markdownify>=1.2.2
great-expectations>=1.15.2
transformers>=4.40.0
torch>=2.0.0
